+++ tests/Type/LazyDefinitionTest.php
@@ -182,4 +182,30 @@ final class LazyDefinitionTest extends TestCaseBase
         self::assertSame($field->name, 'width');
         self::assertInstanceOf(IntType::class, $field->getType());
     }
+
+    public function testLazyRootTypes(): void
+    {
+        $query = new ObjectType([
+            'name' => 'Query',
+            'fields' => [],
+        ]);
+        $mutation = new ObjectType([
+            'name' => 'Mutation',
+            'fields' => [],
+        ]);
+        $subscription = new ObjectType([
+            'name' => 'Subscription',
+            'fields' => [],
+        ]);
+
+        $schema = new Schema([
+            'query' => fn () => $query,
+            'mutation' => fn () => $mutation,
+            'subscription' => fn () => $subscription,
+        ]);
+
+        self::assertSame($schema->getQueryType(), $query);
+        self::assertSame($schema->getMutationType(), $mutation);
+        self::assertSame($schema->getSubscriptionType(), $subscription);
+    }
 }