+++ tests/Type/ValidationTest.php
@@ -375,29 +375,30 @@ final class ValidationTest extends TestCaseBase
     /** @dataProvider rootTypes */
     public function testRejectsASchemaWhoseRootTypeIsAnInputType(string $rootType): void
     {
-        $this->expectRootTypeMustBeObjectTypeNotInputType($rootType);
+        $ucfirstRootType = \ucfirst($rootType);
 
-        BuildSchema::build('
-      input ' . \ucfirst($rootType) . ' {
+        $this->expectRootTypeMustBeObjectTypeNotInputType();
+        BuildSchema::build("
+      input {$ucfirstRootType} {
         test: String
       }
-        ');
+        ");
     }
 
     /** @dataProvider rootTypes */
     public function testRejectsASchemaWhoseNonStandardRootTypeIsAnInputType(string $rootType): void
     {
-        $this->expectRootTypeMustBeObjectTypeNotInputType($rootType);
+        $this->expectRootTypeMustBeObjectTypeNotInputType();
 
-        BuildSchema::build('
+        BuildSchema::build("
       schema {
-        ' . $rootType . ': SomeInputObject
+        {$rootType}: SomeInputObject
       }
 
       input SomeInputObject {
         test: String
       }
-        ');
+        ");
     }
 
     /**
@@ -413,16 +414,14 @@ final class ValidationTest extends TestCaseBase
             }
         ');
 
-        $this->expectRootTypeMustBeObjectTypeNotInputType($rootType);
+        $documentNode = Parser::parse("
+            extend schema {
+              {$rootType}: SomeInputObject
+            }
+        ");
 
-        SchemaExtender::extend(
-            $schema,
-            Parser::parse('
-                extend schema {
-                  ' . $rootType . ': SomeInputObject
-                }
-            ')
-        );
+        $this->expectRootTypeMustBeObjectTypeNotInputType();
+        SchemaExtender::extend($schema, $documentNode);
     }
 
     /** @return iterable<array{string}> */
@@ -433,10 +432,10 @@ final class ValidationTest extends TestCaseBase
         yield ['subscription'];
     }
 
-    private function expectRootTypeMustBeObjectTypeNotInputType(string $rootType): void
+    private function expectRootTypeMustBeObjectTypeNotInputType(): void
     {
-        $this->expectException(\TypeError::class);
-        $this->expectExceptionMessageMatches('/.*GraphQL\\\\Type\\\\SchemaConfig::set' . \ucfirst($rootType) . '.*GraphQL\\\\Type\\\\Definition\\\\ObjectType.*GraphQL\\\\Type\\\\Definition\\\\InputObjectType given.*/');
+        $this->expectException(InvariantViolation::class);
+        $this->expectExceptionMessage("Expected instanceof GraphQL\Type\Definition\ObjectType, a callable that returns such an instance, or null, got: GraphQL\Type\Definition\InputObjectType.");
     }
 
     /** @see it('rejects a Schema whose directives are incorrectly typed') */